Crate oak_utils

Structs

CodegenOptions

Options for building gRPC code.

ExternPath

ProtoOptions

Options for generating Protocol buffer Rust types.

Traits

LogError

Trait for logging error messages

Functions

compile_protos

Build Rust code corresponding to a set of protocol buffer message and service definitions, emitting generated code to crate’s OUT_DIR. For gRPC service definitions, this function generates Oak-specific code that is suitable for use inside an Oak Node (i.e. not code that is suitable for use in a normal application running on the host platform).

compile_protos_with_options

Like compile_protos, but allows for configuring options through ProtoOptions.

generate_grpc_code

Generate gRPC code from Protobuf using tonic library.